requirements.txt
pandas
numpy
seaborn
#matplotlib
#scikit-learn
#catboost
#xgboost
# keras
fitz
Pillow
opencv-python
pytesseract
PyMuPDF
Flask
spacy

-e .